Fixing UI for event stage detail.
