+++ core/displaymodels/src/commonMain/kotlin/com/adammcneilly/pocketleague/core/displaymodels/MatchDetailsByDateDisplayModel.kt
@@ -8,4 +8,19 @@ package com.adammcneilly.pocketleague.core.displaymodels
 data class MatchDetailsByDateDisplayModel(
     val matchesByDate: Map<String, List<MatchDetailDisplayModel>> = emptyMap(),
     val isPlaceholder: Boolean = false,
-)
+) {
+
+    companion object {
+        val placeholder: MatchDetailsByDateDisplayModel
+            get() {
+                val matchesByDate = (1..3).map {
+                    MatchDetailDisplayModel.placeholder
+                }.groupBy(MatchDetailDisplayModel::localDate)
+
+                return MatchDetailsByDateDisplayModel(
+                    matchesByDate = matchesByDate,
+                    isPlaceholder = true,
+                )
+            }
+    }
+}
