+++ shared/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/screens/eventstagedetail/EventStageDetailViewState.kt
@@ -1,6 +1,5 @@
 package com.adammcneilly.pocketleague.shared.screens.eventstagedetail
 
-import com.adammcneilly.pocketleague.core.displaymodels.MatchDetailDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.MatchDetailsByDateDisplayModel
 import com.adammcneilly.pocketleague.core.feature.ScreenState
 import com.adammcneilly.pocketleague.core.models.EventStage
@@ -11,20 +10,8 @@ import com.adammcneilly.pocketleague.core.models.EventStage
 data class EventStageDetailViewState(
     val eventId: String = "",
     val stageId: String = "",
-    val matchesDataState: MatchDetailsByDateDisplayModel? = null,
+    val matchesByDateDisplayModel: MatchDetailsByDateDisplayModel = MatchDetailsByDateDisplayModel.placeholder,
 ) : ScreenState {
 
     override val title: String? = null
-
-    val matchesByDateDisplayModel: MatchDetailsByDateDisplayModel?
-        get() = matchesDataState?.let {
-            val matchesByDate = (1..3).map {
-                MatchDetailDisplayModel.placeholder
-            }.groupBy(MatchDetailDisplayModel::localDate)
-
-            MatchDetailsByDateDisplayModel(
-                matchesByDate = matchesByDate,
-                isPlaceholder = true,
-            )
-        }
 }
