+++ shared/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/screens/eventstagedetail/LoadEventStageDetailEvent.kt
@@ -39,7 +39,7 @@ private fun Events.fetchMatchesForStage(
 
             stateManager.updateScreen(EventStageDetailViewState::class) { currentState ->
                 currentState.copy(
-                    matchesDataState = matchesByDateDisplayModel,
+                    matchesByDateDisplayModel = matchesByDateDisplayModel,
                 )
             }
         }
